+++ src/flagDescriptions.js
@@ -2,7 +2,7 @@ const flagDescriptions = {
     b: {
         title: "Backgrounds",
         max: 1,
-        default: 1, 
+        default: 1,
         desc: {
             1: "In-battle backgrounds will be randomized. Warning: If you are sensitive to bright flashing lights causing epilepsy or seizures, you may not want to enable this mode.",
         }
@@ -10,7 +10,7 @@ const flagDescriptions = {
     c: {
         title: "Character stats",
         max: 1,
-        default: 1, 
+        default: 1,
         desc: {
             1: "Player character's stat increases on each level-up will be randomized.",
         }
@@ -18,7 +18,7 @@ const flagDescriptions = {
     d: {
         title: "Dialogue",
         max: 2,
-        default: 2, 
+        default: 2,
         desc: {
             1: "NPCs with non-vital dialogue will have their lines shuffled.",
             2: "NPCs with non-vital dialogue will have their lines shuffled, and Heavily Armed Pokey will have random special lines.",
@@ -27,16 +27,26 @@ const flagDescriptions = {
     g: {
         title: "Gift box contents",
         max: 2,
-        default: 1, 
+        default: 1,
         desc: {
             1: "If in Ancient Cave, gift boxes will have a logical progression of gear and supplies. In other modes, gift boxes will be replaced with similar content, with a 20% change of being replaced by anything.",
             2: "If in Ancient Cave, gift boxes will have a logical progression of gear and supplies. In other modes, gift boxes will have a 100% change of being replaced by anything.",
         }
     },
+    i: {
+        title: "PSI abilities",
+        max: 2,
+        default: 1,
+        unsafe: 2,
+        desc: {
+            1: "Levels that PSI abilities are learned at will be randomized.",
+            2: "Levels that PSI abilities are learned at, as well as which abilities are learned by which PC, will be randomized.",
+        }
+    },
     m: {
         title: "Enemy stats",
         max: 4,
-        default: 2, 
+        default: 2,
         unsafe: 4,
         desc: {
             1: "Enemy stats and palettes will be randomized.",
@@ -48,7 +58,7 @@ const flagDescriptions = {
     n: {
         title: "NPC sprites",
         max: 3,
-        default: 1, 
+        default: 1,
         unsafe: 2,
         desc: {
             1: "Overworld NPC sprites will be randomized.",
@@ -59,7 +69,7 @@ const flagDescriptions = {
     p: {
         title: "PC sprites",
         max: 2,
-        default: 1, 
+        default: 1,
         unsafe: 2,
         desc: {
             1: "PC sprites will be randomized.",
@@ -69,7 +79,7 @@ const flagDescriptions = {
     s: {
         title: "Shops",
         max: 1,
-        default: 1, 
+        default: 1,
         desc: {
             1: "Shop contents will be randomized with similar contents.",
         }
@@ -88,7 +98,7 @@ const flagDescriptions = {
     devmode: {
         title: "Developer mode",
         max: 2,
-        default: 0, 
+        default: 0,
         desc: {
             0: "Disabled.",
             1: "Extra info in spoiler file and other development-oriented changes.",
@@ -98,7 +108,7 @@ const flagDescriptions = {
     easymodo: {
         title: "Easy mode",
         max: 1,
-        default: 0, 
+        default: 0,
         desc: {
             0: "Disabled.",
             1: "Lvl 99, protective gear, enemy HP 1, no spawn plates.",
@@ -107,7 +117,7 @@ const flagDescriptions = {
     giygastest: {
         title: "Giygas test mode",
         max: 1,
-        default: 0, 
+        default: 0,
         desc: {
             0: "Disabled.",
             1: "Ness's room connected to Giygas.",